CONF State Addition
Adds a configuration state in which the sensor starts a soft WiFi access point. Previous code is now part of the MEAS state.

The transition to the CONF state occurs when the sensor starts/resets and the pin connected to the CONF pushbutton is closed.

Takes also the opportunity to change the UART speed to match the default of PlateformIO