Receive Data from HTTP POST
Updates the WiFi SSID and password with the HTTP POST data. For now, the EEPROM is not updated with the new values. In consequence, the new SSID and password will not survive a chip reset.

Removes some notice (*N*) on the serial interface. They were making debugging difficult at they were really verbose.